crypto: move P256 derivation points into primitives
