+++ program-libs/hasher/src/lib.rs
@@ -29,6 +29,7 @@ pub mod bigint;
 pub mod errors;
 pub mod hash_chain;
 pub mod keccak;
+pub mod p256;
 pub mod poseidon;
 pub mod primitives;
 pub mod sha256;
