+++ program-libs/hasher/src/p256.rs
@@ -1,0 +1,21 @@
+pub const P_CONST_SEC1: [u8; 33] = [
+    0x03, 0x0e, 0x4d, 0xf9, 0x46, 0xbc, 0xe1, 0x4b, 0x95, 0x29, 0x2f, 0x13, 0xe1, 0x33, 0xd2, 0xb0,
+    0xc6, 0x4e, 0x89, 0x8b, 0x56, 0x44, 0xf6, 0x20, 0xa5, 0xbe, 0xd2, 0x5a, 0x06, 0x1a, 0x42, 0xfc,
+    0xdb,
+];
+
+pub const P_DERIVE_SEC1: [u8; 33] = [
+    0x03, 0x9e, 0xf1, 0x65, 0x92, 0x42, 0x9d, 0xa1, 0x40, 0x3e, 0xaa, 0x29, 0x05, 0x8e, 0xb7, 0xd9,
+    0xd5, 0xad, 0x15, 0xa2, 0xea, 0x55, 0x71, 0x74, 0xf7, 0xb0, 0x1f, 0xf7, 0xfe, 0x48, 0x4e, 0xee,
+    0xaf,
+];
+
+pub const P_PDA_SEC1: [u8; 33] = [
+    0x03, 0x8a, 0x31, 0xd5, 0x3c, 0x5a, 0xd2, 0x0d, 0x1c, 0xd7, 0xee, 0x1f, 0xbb, 0x99, 0x27, 0xbd,
+    0x0c, 0xdf, 0xb6, 0x1b, 0x1b, 0x89, 0xf6, 0xb2, 0xc5, 0xa9, 0x4a, 0x5f, 0x08, 0x1f, 0xe1, 0x6b,
+    0x5b,
+];
+
+pub fn is_reserved_derivation_point(key: &[u8; 33]) -> bool {
+    key == &P_CONST_SEC1 || key == &P_DERIVE_SEC1 || key == &P_PDA_SEC1
+}
